move upsample2d and downsample2d to utils.
